move test cleanup to post action

+++ src/com/oracle/fmwk8s/test/Test.groovy
@@ -1,6 +1,7 @@
 package com.oracle.fmwk8s.test
 
 import com.oracle.fmwk8s.common.Common
+import com.oracle.fmwk8s.common.EnvironmentSetup
 import com.oracle.fmwk8s.common.Log
 import com.oracle.fmwk8s.env.Domain
 
@@ -68,22 +69,24 @@ class Test extends Common {
     }
 
     static cleanup() {
-        try {
-            Log.info("begin cleanup test resources.")
+        if (!EnvironmentSetup.isWaiting) {
+            try {
+                Log.info("begin cleanup test resources.")
 
-            script.sh label: "cleanup test pod",
-                    script: "kubectl delete -f kubernetes/framework/test/${testId}/fmwk8s-${testId}-test-pod.yaml -n ${Domain.domainNamespace} --grace-period=0 --force --cascade"
-            sleep 30
-            script.sh label: "cleanup test pv/pvc",
-                    script: "kubectl delete fmwk8s-tests-pvc-${Common.runId} -n ${Domain.domainNamespace} --grace-period=0 --force --cascade && \
-                             sleep 30 && \
-                             kubectl delete fmwk8s-tests-pv-${Common.runId} -n ${Domain.domainNamespace} --grace-period=0 --force --cascade"
+                script.sh label: "cleanup test pod",
+                        script: "kubectl delete -f kubernetes/framework/test/${testId}/fmwk8s-${testId}-test-pod.yaml -n ${Domain.domainNamespace} --grace-period=0 --force --cascade"
+                sleep 30
+                script.sh label: "cleanup test pv/pvc",
+                        script: "kubectl delete fmwk8s-tests-pvc-${Common.runId} -n ${Domain.domainNamespace} --grace-period=0 --force --cascade && \
+                                 sleep 30 && \
+                                 kubectl delete fmwk8s-tests-pv-${Common.runId} -n ${Domain.domainNamespace} --grace-period=0 --force --cascade"
 
-            Log.info("cleanup test resources success.")
-        }
-        catch (exc) {
-            Log.error("cleanup test resources failed.")
-            exc.printStackTrace()
+                Log.info("cleanup test resources success.")
+            }
+            catch (exc) {
+                Log.error("cleanup test resources failed.")
+                exc.printStackTrace()
+            }
         }
     }
 }
